feat(cli): move lifecycle commands under repo
