+++ packages/cli/src/commands/setup.ts
@@ -30,7 +30,7 @@ import { validateStateMapping } from "../mapping/smart-defaults.js";
 import {
   promptProjectRegistrationOptions,
   renderProjectRegistrationSummary,
-} from "./project.js";
+} from "./setup-project-registration.js";
 
 const KNOWN_REQUIRED_SCOPES = ["repo", "read:org", "project"] as const;
